move modal youtube logic into the columns
