+++ blocks/columns/columns.css
@@ -70,7 +70,7 @@ main .section.columns-container > div{
 }
 
 /* Block styleing */
-.columns.homepage.block.columns-2-cols > div {
+.columns.homepage.columns-2-cols > div {
   position: relative;
   z-index: 1;
   display: flex;
@@ -82,32 +82,32 @@ main .section.columns-container > div{
   box-sizing: border-box;
 }
 
-.columns.homepage.block.columns-2-cols > div > .text-content {
+.columns.homepage.columns-2-cols .text-content {
   order: 1;
   width: 100%;
 }
 
-.columns.homepage.block.columns-2-cols > div > .text-content img{
+.columns.homepage.columns-2-cols .text-content img{
   width: 200%;
   max-width: 400px;
   height: auto;
   display:none;
 }
 
-.columns.homepage.block.columns-2-cols > div > .columns-img-col {
+.columns.homepage.columns-2-cols .columns-img-col {
   order: 3;
   width: 100%;
 }
 
-.columns.homepage.block.columns-2-cols > div > .columns-img-col img {
+.columns.homepage.columns-2-cols .columns-img-col img {
   max-width: 100%;
   height: auto;
 }
 
 /* Full-width colored background */
-.columns.homepage.block.columns-2-cols > div.bg-color-1::before,
-.columns.homepage.block.columns-2-cols > div.bg-color-2::before,
-.columns.homepage.block.columns-2-cols > div.bg-color-3::before {
+.columns.homepage.columns-2-cols > div.bg-color-1::before,
+.columns.homepage.columns-2-cols > div.bg-color-2::before,
+.columns.homepage.columns-2-cols > div.bg-color-3::before {
   content: '';
   position: absolute;
   top: 0;
@@ -128,6 +128,91 @@ main .section.columns-container > div{
 
 .section.section-pale-blue-lighter .columns.homepage.block .button:active {
   background-color: var(--button-blue-active-color);
+}
+
+.section.multiple-columns.columns-container {
+  display: flex;
+  justify-content: center;
+  align-items: stretch;
+}
+
+.section.multiple-columns.columns-container .columns-wrapper {
+  flex: 1;
+  max-width: 460px;
+  box-sizing: border-box;
+  border-radius: 10px;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #fff;
+  margin: 2rem 0;
+}
+
+.section.multiple-columns.columns-container .columns {
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+}
+
+.section.multiple-columns.columns-container .columns.biege {
+  background-color: #FAF9F7;
+}
+
+.section.multiple-columns.columns-container .columns.warm-gray {
+  background-color: #F6F3F0;
+}
+
+.section.multiple-columns.columns-container .columns-wrapper img {
+  max-width: 100%;
+  height: auto;
+  border-radius: 10px;
+}
+
+.section.multiple-columns.columns-container .columns-wrapper h2, 
+.section.multiple-columns.columns-container .columns-wrapper p {
+  text-align: center;
+  padding: 0.5em 1.2em;
+  font-family: var(--font-family);
+}
+
+.section.multiple-columns.columns-container .columns-wrapper h2 {
+  font-size: 1.5em;
+  margin-bottom: 0;
+}
+
+.section.multiple-columns.columns-container .columns-wrapper p {
+  font-size: 0.8em;
+}
+
+.section.multiple-columns.columns-container .columns-wrapper .button-container {
+  margin-top: auto;
+  cursor: pointer;
+}
+
+.section.multiple-columns.columns-container .columns .youtube-thumbnail {
+  border-radius: 20px;
+  padding: 1rem;
+  width: 90%;
+  height: 200px;
+  object-fit: cover;
+  margin: 0 auto;
+  display: block;
+}
+
+.section.multiple-columns.columns-container .columns .columns-img-col picture {
+  max-height: 200px;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+}
+
+.section.multiple-columns.columns-container .columns .columns-img-col picture img {
+  max-height: 100%;
+  width: auto;
 }
 
 /* Mobile-specific adjustments */
@@ -345,5 +430,16 @@ main .section.columns-container > div{
 
   .columns.footer.block.columns-3-cols .text-content p {
     padding: 0;
+  }
+  
+  .section.multiple-columns.columns-container {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .section.multiple-columns.columns-container .columns-wrapper {
+    max-width: 100%;
+    margin:0;
+    padding:0;
   }
 }
